Feature: survey schema summary download in dashboard page
