+++ webapp/views/App/views/Home/Dashboard/SurveyInfo/SurveyInfo.js
@@ -10,15 +10,14 @@ import { appModuleUri, homeModules } from '@webapp/app/appModules'
 import { useI18n } from '@webapp/store/system'
 import { SurveyActions, useSurveyInfo } from '@webapp/store/survey'
 import { useAuthCanEditSurvey } from '@webapp/store/user'
+import { useAuthCanExportSurvey } from '@webapp/store/user/hooks'
 import { TestId } from '@webapp/utils/testId'
 
+import { useConfirmDelete } from '@webapp/components/hooks'
 import Header from '@webapp/components/header'
 import ButtonPublishSurvey from '@webapp/components/buttonPublishSurvey'
-import { Button, ButtonDownload } from '@webapp/components'
-
-import { useAuthCanExportSurvey } from '@webapp/store/user/hooks'
-import { ButtonMenu } from '@webapp/components/buttons'
-import { useConfirmDelete } from '@webapp/components/hooks'
+import { Button, ButtonDownload, ButtonMenu } from '@webapp/components'
+import SurveySchemaSummaryDownloadButton from '@webapp/components/survey/SurveySchemaSummaryDownloadButton'
 
 const SurveyInfo = () => {
   const i18n = useI18n()
@@ -72,47 +71,48 @@ const SurveyInfo = () => {
             <span className={`icon icon-${canEditSurvey ? 'pencil2' : 'eye'} icon-12px icon-left`} />
             {i18n.t(canEditSurvey ? 'homeView.surveyInfo.editInfo' : 'homeView.surveyInfo.viewInfo')}
           </Link>
-
           {canEditSurvey && <ButtonPublishSurvey className="btn-transparent" disabled={!Survey.isDraft(surveyInfo)} />}
+          <ButtonMenu
+            testId={TestId.dashboard.surveyExportBtn}
+            className="btn-menu-export"
+            buttonClassName="btn-transparent"
+            iconClassName="icon-download2 icon-14px"
+            title="common.export"
+            popupComponent={
+              <>
+                <SurveySchemaSummaryDownloadButton className="btn-transparent" />
 
-          {canExportSurvey && (
-            <ButtonMenu
-              testId={TestId.dashboard.surveyExportBtn}
-              className="btn-menu-export"
-              buttonClassName="btn-transparent"
-              iconClassName="icon-download2 icon-14px"
-              title="common.export"
-              popupComponent={
-                <>
-                  <ButtonDownload
-                    testId={TestId.dashboard.surveyExportOnlySurveyBtn}
-                    className="btn-transparent"
-                    onClick={() => dispatch(SurveyActions.exportSurvey())}
-                    label="common.export"
-                  />
-                  {!Survey.isTemplate(surveyInfo) && (
-                    <>
-                      <ButtonDownload
-                        testId={TestId.dashboard.surveyExportWithDataBtn}
-                        className="btn-transparent"
-                        onClick={() => dispatch(SurveyActions.exportSurvey({ includeData: true }))}
-                        label="homeView.dashboard.exportWithData"
-                      />
-                      <ButtonDownload
-                        testId={TestId.dashboard.surveyExportWithDataNoActivityLogBtn}
-                        className="btn-transparent"
-                        onClick={() =>
-                          dispatch(SurveyActions.exportSurvey({ includeData: true, includeActivityLog: false }))
-                        }
-                        label="homeView.dashboard.exportWithDataNoActivityLog"
-                      />
-                    </>
-                  )}
-                </>
-              }
-            />
-          )}
-
+                {canExportSurvey && (
+                  <>
+                    <ButtonDownload
+                      testId={TestId.dashboard.surveyExportOnlySurveyBtn}
+                      className="btn-transparent"
+                      onClick={() => dispatch(SurveyActions.exportSurvey())}
+                      label="common.export"
+                    />
+                    {!Survey.isTemplate(surveyInfo) && (
+                      <>
+                        <ButtonDownload
+                          testId={TestId.dashboard.surveyExportWithDataBtn}
+                          className="btn-transparent"
+                          onClick={() => dispatch(SurveyActions.exportSurvey({ includeData: true }))}
+                          label="homeView.dashboard.exportWithData"
+                        />
+                        <ButtonDownload
+                          testId={TestId.dashboard.surveyExportWithDataNoActivityLogBtn}
+                          className="btn-transparent"
+                          onClick={() =>
+                            dispatch(SurveyActions.exportSurvey({ includeData: true, includeActivityLog: false }))
+                          }
+                          label="homeView.dashboard.exportWithDataNoActivityLog"
+                        />
+                      </>
+                    )}
+                  </>
+                )}
+              </>
+            }
+          />
           {canEditSurvey && (
             <Button
               className="btn-s btn-transparent"
@@ -122,7 +122,6 @@ const SurveyInfo = () => {
               label="common.delete"
             />
           )}
-
           {canEditSurvey && Survey.isFromCollect(surveyInfo) && Survey.hasCollectReportIssues(surveyInfo) && (
             <Link
               data-testid={TestId.dashboard.collectReportBtn}
